+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/generators/ContributedInterfaceSupertypeGenerator.kt
@@ -15,6 +15,7 @@ import dev.zacsweers.metro.compiler.fir.metroFirBuiltIns
 import dev.zacsweers.metro.compiler.fir.predicates
 import dev.zacsweers.metro.compiler.fir.qualifierAnnotation
 import dev.zacsweers.metro.compiler.fir.rankValue
+import dev.zacsweers.metro.compiler.fir.resolveClassId
 import dev.zacsweers.metro.compiler.fir.resolvedAdditionalScopesClassIds
 import dev.zacsweers.metro.compiler.fir.resolvedBindingArgument
 import dev.zacsweers.metro.compiler.fir.resolvedExcludedClassIds
@@ -147,14 +148,11 @@ internal class ContributedInterfaceSupertypeGenerator(session: FirSession) :
     return buildMap {
       for (originClass in contributingClasses) {
         if (originClass.isAnnotatedWithAny(session, session.classIds.bindingContainerAnnotations)) {
-          val scopeId =
-            originClass
-              .annotationsIn(session, session.classIds.contributesToAnnotations)
-              .single()
-              .resolvedScopeClassId(typeResolver)
-          if (scopeId == scopeClassId) {
-            put(originClass.classId, true)
-          }
+          val hasMatchingScope =
+            originClass.annotationsIn(session, session.classIds.contributesToAnnotations).any {
+              it.resolvedScopeClassId(typeResolver) == scopeClassId
+            }
+          put(originClass.classId, hasMatchingScope)
           continue
         }
 
@@ -331,7 +329,8 @@ internal class ContributedInterfaceSupertypeGenerator(session: FirSession) :
           typeResolverFactory.create(contributingType) ?: return@flatMap emptySequence()
 
         contributingType
-          .annotationsIn(session, session.classIds.allContributesAnnotations)
+          .annotationsIn(session, session.classIds.allContributesAnnotationsWithContainers)
+          .filter { it.scopeArgument()?.resolveClassId(localTypeResolver) in scopes }
           .flatMap { annotation -> annotation.resolvedReplacedClassIds(localTypeResolver) }
       }
       .distinct()
@@ -348,7 +347,7 @@ internal class ContributedInterfaceSupertypeGenerator(session: FirSession) :
 
     if (session.metroFirBuiltIns.options.enableDaggerAnvilInterop) {
       val unmatchedRankReplacements = mutableSetOf<ClassId>()
-      val pendingRankReplacements = processRankBasedReplacements(contributions, typeResolver)
+      val pendingRankReplacements = processRankBasedReplacements(scopes, contributions, typeResolver)
 
       pendingRankReplacements.distinct().forEach { replacedClassId ->
         val removed = contributions.remove(replacedClassId)
@@ -372,13 +371,14 @@ internal class ContributedInterfaceSupertypeGenerator(session: FirSession) :
   }
 
   /**
-   * This provides ContributesBinding.rank interop for users migrating from Dagger-Anvil to make the
-   * migration to Metro more feasible.
+   * This provides `ContributesBinding.rank` interop for users migrating from Dagger-Anvil to make
+   * the migration to Metro more feasible.
    *
    * @return The bindings which have been outranked and should not be included in the merged graph.
    */
   private fun processRankBasedReplacements(
-    contributions: TreeMap<ClassId, ConeKotlinType>,
+    allScopes: Set<ClassId>,
+    contributions: Map<ClassId, ConeKotlinType>,
     typeResolver: TypeResolveService,
   ): Set<ClassId> {
     val pendingRankReplacements = mutableSetOf<ClassId>()
@@ -390,9 +390,11 @@ internal class ContributedInterfaceSupertypeGenerator(session: FirSession) :
         .flatMap { contributingType ->
           contributingType
             .annotationsIn(session, session.classIds.contributesBindingAnnotations)
-            // TODO Can enforce non-null boundTypes here once type arguments are saved to metadata
-            // https://youtrack.jetbrains.com/issue/KT-76954/Some-type-arguments-are-not-saved-to-metadata-in-FIR
             .mapNotNull { annotation ->
+
+              val scope = annotation.resolvedScopeClassId(typeResolver) ?: return@mapNotNull null
+              if (scope !in allScopes) return@mapNotNull null
+
               val explicitBindingMissingMetadata =
                 annotation.argumentAsOrNull<FirAnnotation>(Symbols.Names.binding, index = 1)
 
@@ -414,16 +416,18 @@ internal class ContributedInterfaceSupertypeGenerator(session: FirSession) :
                   } ?: contributingType.implicitBoundType(typeResolver)
 
                 ContributedBinding(
-                  contributingType,
-                  FirTypeKey(
-                    boundType,
-                    contributingType.qualifierAnnotation(session, typeResolver),
-                  ),
-                  annotation.rankValue(),
+                  contributingType = contributingType,
+                  typeKey =
+                    FirTypeKey(
+                      boundType,
+                      contributingType.qualifierAnnotation(session, typeResolver),
+                    ),
+                  rank = annotation.rankValue(),
                 )
               }
             }
         }
+
     val bindingGroups =
       rankedBindings
         .groupBy { binding -> binding.typeKey }