+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/DependencyGraphNodeCache.kt
@@ -171,6 +171,7 @@ internal class DependencyGraphNodeCache(
           val ctorParams = ctor.parameters()
           populateBindingContainerFields(ctorParams)
           DependencyGraphNode.Creator.Constructor(
+            graphDeclaration,
             graphDeclaration.primaryConstructor!!,
             ctorParams,
             bindingContainerFields,
@@ -418,10 +419,10 @@ internal class DependencyGraphNodeCache(
           it.classType.rawTypeOrNull()?.classId
         }
 
-      dependencyGraphAnno?.scopeClassOrNull()?.let { scope ->
+      for (scope in aggregationScopes) {
         bindingContainers +=
           contributionData
-            .getBindingContainerContributions(scope.classIdOrFail)
+            .getBindingContainerContributions(scope)
             .mapNotNull { bindingContainerTransformer.findContainer(it) }
             .filterNot { it.ir.classId in excludes }
             .onEach { container ->