+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/DependencyGraphTransformer.kt
@@ -154,9 +154,10 @@ internal class DependencyGraphTransformer(
       return
     }
 
+    val tag = dependencyGraphDeclaration.kotlinFqName.shortName().asString()
     parentTracer.traceNested(
-      "Transform dependency graph",
-      dependencyGraphDeclaration.kotlinFqName.shortName().asString(),
+      "[$tag] Transform dependency graph",
+      tag,
     ) { tracer ->
       transformDependencyGraph(
         graphClassId,
@@ -214,11 +215,11 @@ internal class DependencyGraphTransformer(
               for ((declaration, message) in errors) {
                 // TODO in kotlin 2.2.20 we can just use the reporter
                 val toReport =
-                  declaration?.takeIf { it.fileOrNull != null } ?: dependencyGraphDeclaration
-                if (toReport.fileOrNull != null) {
-                  diagnosticReporter
-                    .at(declaration ?: dependencyGraphDeclaration)
-                    .report(MetroIrErrors.METRO_ERROR, message)
+                  declaration?.takeIf {
+                    it.fileOrNull != null && it.origin != Origins.ContributedGraph
+                  } ?: dependencyGraphDeclaration
+                if (toReport.fileOrNull != null && toReport.origin != Origins.ContributedGraph) {
+                  diagnosticReporter.at(toReport).report(MetroIrErrors.METRO_ERROR, message)
                 } else {
                   messageCollector.report(
                     CompilerMessageSeverity.ERROR,